refactor: `numberChecker` 추상 클래스를 활용해 숫자 공통 검증 상속
