+++ src/main/java/lotto/domain/lottery/LottoService.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottery;
 
 import camp.nextstep.edu.missionutils.Randoms;
